Remove unneeded ellipis, and other small string changes
The Haiku Interface Guide (which is currently MIA), says to not
use ellipsis even when the menu item opens a window, if the label
of the menu is actually done when it's invoked. Hmm... not easy to
explain. Example:

Ellipsis OK: "New project..." opens a window to set up a new project.
              After the window has opened, you're not done.

Ellipsis not OK: "ASCII table" opens the actual ASCII table. you're done.